Backend/requirements.txt
python-dotenv
flask
flask-cors
pylint